Move auth gate to view model
